refactor: created a single form for both components
